исправил все замечания

+++ src/Strategy/PaymentStrategy.java
@@ -1,5 +1,7 @@
 package Strategy;
 
+import java.math.BigDecimal;
+
 public interface PaymentStrategy {
-    void pay(int amount);
+    void pay(BigDecimal amount);
 }
